pyproject.toml
[tool.poetry]
name = "to-do-app"
version = "0.1.0"
description = ""
authors = ["Joanna Cheng <joanna.cheng@softwire.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.24.0"
gunicorn = "^20.0.4"
lxml = "^4.6.2"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
selenium = "^3.141.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[virtualenvs]
create = true
in-project = false
